requirements.txt
numpy==1.23.4
pandas==1.4.4
Pillow==10.2.0
plotly==5.19.0
scikit_learn==1.4.1.post1
torch==1.13.0
torchmetrics==1.3.1
torchvision==0.14.0
tqdm==4.64.0
wandb==0.16.3
